server/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0

# Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process
# .\venv\Scripts\activate
